### 2024-01-19 #6247 @editorialbot
[PRE REVIEW]: NCDatasets.jl: a Julia package for manipulating netCDF data sets
**Submitting author:** @Alexander-Barth (<a href="http://orcid.org/0000-0003-2952-5997">Alexander Barth</a>)
**Repository:** https://github.com/Alexander-Barth/NCDatasets.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.14
