### 2024-01-20 #6247 @Alexander-Barth
@kthyng Thank you for letting me know. It seems to be fixed now (replacing tabs by spaces).

[I tested the submission before using the docker image (`docker run --rm  --volume $PWD/paper:/data --user $(id -u):$(id -g)  --env JOURNAL=joss openjournals/inara`), but this did not show the error.]
