+++ src/NCDatasets.jl
@@ -47,7 +47,7 @@ import CommonDataModel: AbstractDataset, AbstractVariable,
     SubDataset,
     @select, select, Near, coordinate_value, coordinate_names, split_by_and,
     chunking, deflate, checksum,
-    _experimental_missing_value
+    maskingvalue
 
 
 import DiskArrays
