+++ test/test_copyvar.jl
@@ -2,22 +2,22 @@ using Dates
 using NCDatasets
 using Test
 
-var = 10.0:10.0:40.0
-tax = DateTime(2001,1,1) .+ Day.(Int.(var))
+datavar = 10.0:10.0:40.0
+tax = DateTime(2001,1,1) .+ Day.(Int.(datavar))
 fname = tempname()
 fname2 = tempname()
 ds = NCDataset(fname, "c")
 defDim(ds, "time", Inf) # "unlimited"
 defVar(ds, "time", tax, ("time",))
-defVar(ds, "var", var, ("time",),deflatelevel=9)
+defVar(ds, "datavar", datavar, ("time",),deflatelevel=9)
 close(ds)
 
 NCDataset(fname, "r") do ds
     time = ds["time"]
-    var = ds["var"]
+    datavar = ds["datavar"]
     NCDataset(fname2, "c") do ds2
         defVar(ds2, "time", time, ("time",))
-        defVar(ds2, "var",  var,  ("time",))
+        defVar(ds2, "datavar",  datavar,  ("time",))
         @test "time" in unlimited(ds)
   end
 end
@@ -25,9 +25,9 @@ end
 NCDataset(fname, "r") do ds
     NCDataset(fname2, "c") do ds2
         defVar(ds2, ds["time"])
-        defVar(ds2, ds["var"])
+        defVar(ds2, ds["datavar"])
         @test "time" in unlimited(ds)
-        isshuffled,isdeflated,deflatelevel = deflate(ds["var"])
+        isshuffled,isdeflated,deflatelevel = deflate(ds["datavar"])
         @test deflatelevel == 9
   end
 end
