+++ .github/workflows/ci.yml
@@ -11,6 +11,7 @@ jobs:
       matrix:
         version:
           - '1.6'
+          - '1.7'          
           - '1'
           #- '~1.8.0-0'
           - 'nightly'
