+++ .github/workflows/ci.yml
@@ -11,7 +11,9 @@ jobs:
       matrix:
         version:
           - '1.6'
-          - '1.7'          
+          - '1.7'
+          - '1.8'
+          - '1.9'
           - '1'
           #- '~1.8.0-0'
           - 'nightly'
