+++ v0.14.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"dataset/#Datasets","page":"Datasets","title":"Datasets","text":"","category":"section"},{"location":"dataset/","page":"Datasets","title":"Datasets","text":"This page is about loading/writing, examining and operating directly on entire NetCDF datasets. For functions regarding the variables stored in them, see the Variables page.","category":"page"},{"location":"dataset/","page":"Datasets","title":"Datasets","text":"Both variables and datasets share the functionality of the Attributes section.","category":"page"},{"location":"dataset/","page":"Datasets","title":"Datasets","text":"NCDataset","category":"page"},{"location":"dataset/#NCDatasets.NCDataset","page":"Datasets","title":"NCDatasets.NCDataset","text":"NCDataset(filename::AbstractString, mode = \"r\";\n          format::Symbol = :netcdf4,\n          share::Bool = false,\n          diskless::Bool = false,\n          persist::Bool = false,\n          memory::Union{Vector{UInt8},Nothing} = nothing,\n          attrib = [])\n\nLoad, create, or even overwrite a NetCDF file at filename, depending on mode\n\n\"r\" (default) : open an existing netCDF file or OPeNDAP URL  in read-only mode.\n\"c\" : create a new NetCDF file at filename (an existing file with the same name will be overwritten).\n\"a\" : open filename into append mode (i.e. existing data in the netCDF file is not overwritten and a variable can be added).\n\nIf share is true, the NC_SHARE flag is set allowing to have multiple processes to read the file and one writer process. Likewise setting diskless or persist to true will enable the flags NC_DISKLESS or NC_PERSIST flag. More information is available in the NetCDF C-API.\n\nNotice that this does not close the dataset, use close on the result (or see below the do-block).\n\nThe optional parameter attrib is an iterable of attribute name and attribute value pairs, for example a Dict, DataStructures.OrderedDict or simply a vector of pairs (see example below).\n\nSupported format values:\n\n:netcdf4 (default): HDF5-based NetCDF format.\n:netcdf4_classic: Only netCDF 3 compatible API features will be used.\n:netcdf3_classic: classic netCDF format supporting only files smaller than 2GB.\n:netcdf3_64bit_offset: improved netCDF format supporting files larger than 2GB.\n:netcdf5_64bit_data: improved netCDF format supporting 64-bit integer data types.\n\nFiles can also be open and automatically closed with a do block.\n\nNCDataset(\"file.nc\") do ds\n    data = ds[\"temperature\"][:,:]\nend\n\nHere is an attribute example:\n\nusing DataStructures\nNCDataset(\"file.nc\", \"c\", attrib = OrderedDict(\"title\" => \"my first netCDF file\")) do ds\n   defVar(ds,\"temp\",[10.,20.,30.],(\"time\",))\nend;\n\nThe NetCDF dataset can also be a memory as a vector of bytes. A non-empty string a filename is still required, for example:\n\nusing NCDataset, HTTP\nresp = HTTP.get(\"https://www.unidata.ucar.edu/software/netcdf/examples/ECMWF_ERA-40_subset.nc\")\nds = NCDataset(\"some_string\",\"r\",memory = resp.body)\ntotal_precipitation = ds[\"tp\"][:,:,:]\nclose(ds)\n\nDataset is an alias of NCDataset.\n\n\n\n\n\nmfds = NCDataset(fnames, mode = \"r\"; aggdim = nothing, deferopen = true,\n              isnewdim = false,\n              constvars = [])\n\nOpens a multi-file dataset in read-only \"r\" or append mode \"a\". fnames is a vector of file names.\n\nVariables are aggregated over the first unlimited dimensi
[…]
