+++ appveyor.yml
@@ -1,7 +1,6 @@
 environment:
   matrix:
-  - julia_version: 1
-  - julia_version: nightly
+  - julia_version: 1.10
 
 platform:
   - x86 # 32-bit
