### 2024-02-09 142ee13
simplify nc_put_att
