+++ versions.js
@@ -14,5 +14,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.14.1";
+var DOCUMENTER_NEWEST = "v0.14.2";
 var DOCUMENTER_STABLE = "stable";
