+++ Project.toml
@@ -19,7 +19,7 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 CFTime = "0.1.1"
 CommonDataModel = "0.3.4"
 DataStructures = "0.17, 0.18"
-DiskArrays = "0.3.22"
+DiskArrays = "0.4"
 NetCDF_jll = "=400.701.400, =400.702.400, =400.902.5, =400.902.208, =400.902.209"
 NetworkOptions = "1.2"
 julia = "1.6"
