### 2024-02-17 #6247 @majensen
@lanari - would you be willing to review [this work](#6247) for [JOSS](https://joss.theoj.org)? I would appreciate your expertise - thanks --@majensen
