### 2024-02-17 #6247 @majensen
Hello @facusapienza21 - would you be willing to review [this work](#6247) for [JOSS](https://joss.theoj.org)? Your expertise on this would be much appreciated! Thanks --@majensen
