+++ src/netcdf_c.jl
@@ -202,12 +202,6 @@ const NC_FILL_INT64  = Int64(-9223372036854775806)
 const NC_FILL_UINT64 = UInt64(18446744073709551614)
 const NC_FILL_STRING = ""
 
-# deprecated by e.g. fillvalue(Int8)
-#export NC_FILL_BYTE, NC_FILL_CHAR, NC_FILL_SHORT, NC_FILL_INT, NC_FILL_FLOAT,
-#    NC_FILL_DOUBLE, NC_FILL_UBYTE, NC_FILL_USHORT, NC_FILL_UINT, NC_FILL_INT64,
-#    NC_FILL_UINT64, NC_FILL_STRING
-
-
 const nc_type = Cint
 
 # type is immutable to ensure that it has the memory same layout
