### 2024-03-01 475a600
CompatHelper: bump compat for "DiskArrays" to "0.4"
