### 2024-03-04 0284393
allow CartesianIndex and CartesianIndices with load! (issue #250)
Co-authored-by: Haakon Ludvig Langeland Ervik <45243236+haakon-e@users.noreply.github.com>
