### 2024-03-04 0335d26
use to_indices rather than own normalizeindexes for index normalization (#250)
Co-authored-by: Haakon Ludvig Langeland Ervik <45243236+haakon-e@users.noreply.github.com>
