### 2024-03-17 #6247 @majensen
@eviatarbach - would you be willing to review [this work](#6247) for [JOSS](https://joss.theoj.org)? We could use your expertise. Thanks for considering --@majensen
