### 2024-03-17 #6247 @majensen
@boriskaus - [JOSS](https://joss.theoj.org) could use your expertise to review [this work](#6247). Would you have time to do so? Thanks very much --@majensen
