### 2024-03-19 #6504 @Alexander-Barth
The editorialbot did not detect the MIT license of the package. I just change the markdown license to a plain-text license which seems to be more in line with recently published joss paper.
