### 2024-03-20 #6504 @boriskaus
@Alexander-Barth: I was checking your readme, and while it all works, I think it could be improved by changing the order of the sections in README by first creating a netCDF file `/tmp/test.nc`, which is then loaded in a next step.
If a new user works through the current README, the first step involves using `ds = NCDataset("file.nc")`, which errors as `file.nc` is not present.
