### 2024-03-20 2fe4efb
Reorder examples in README.md (thanks @boriskaus)
