+++ README.md
@@ -35,11 +35,59 @@ Pkg.add("NCDatasets")
 
 This Manual is a quick introduction in using NCDatasets.jl. For more details you can read the [stable](https://alexander-barth.github.io/NCDatasets.jl/stable/) or [latest](https://alexander-barth.github.io/NCDatasets.jl/latest/) documentation.
 
+* [Create a netCDF file](#create-a-netcdf-file)
 * [Explore the content of a netCDF file](#explore-the-content-of-a-netcdf-file)
 * [Load a netCDF file](#load-a-netcdf-file)
-* [Create a netCDF file](#create-a-netcdf-file)
 * [Edit an existing netCDF file](#edit-an-existing-netcdf-file)
 
+## Create a netCDF file
+
+The following gives an example of how to create a netCDF file by defining dimensions, variables and attributes.
+
+```julia
+using NCDatasets
+using DataStructures: OrderedDict
+# This creates a new NetCDF file called file.nc.
+# The mode "c" stands for creating a new file (clobber)
+ds = NCDataset("file.nc","c")
+
+# Define the dimension "lon" and "lat" with the size 100 and 110 resp.
+defDim(ds,"lon",100)
+defDim(ds,"lat",110)
+
+# Define a global attribute
+ds.attrib["title"] = "this is a test file"
+
+# Define the variables temperature with the attribute units
+v = defVar(ds,"temperature",Float32,("lon","lat"), attrib = OrderedDict(
+    "units" => "degree Celsius",
+    "scale_factor" => 10,
+))
+
+# add additional attributes
+v.attrib["comments"] = "this is a string attribute with Unicode Ω ∈ ∑ ∫ f(x) dx"
+
+# Generate some example data
+data = [Float32(i+j) for i = 1:100, j = 1:110];
+
+# write a single column
+v[:,1] = data[:,1];
+
+# write a the complete data set
+v[:,:] = data;
+
+close(ds)
+```
+
+It is also possible to create the dimensions, the define the variable and set its value with a single call to `defVar`:
+
+```julia
+using NCDatasets
+ds = NCDataset("/tmp/test2.nc","c")
+data = [Float32(i+j) for i = 1:100, j = 1:110]
+v = defVar(ds,"temperature",data,("lon","lat"))
+close(ds)
+```
 ## Explore the content of a netCDF file
 
 Before reading the data from a netCDF file, it is often useful to explore the list of variables and attributes defined in it.
@@ -63,24 +111,28 @@ while to get the global attributes you can do:
 ```julia
 ds.attrib
 ```
-which produces a listing like:
+
+`NCDataset("file.nc")` produces a listing like:
 
 ```
 Dataset: file.nc
 Group: /
 
 Dimensions
-   time = 115
+   lon = 100
+   lat = 110
 
 Variables
-  time   (115)
-    Datatype:    Float64
-    Dimensions:  time
+  temperature   (100 × 110)
+    Datatype:    Float32 (Float32)
+    Dimensions:  lon × lat
     Attributes:
-     calendar             = gregorian
-     standard_name        = time
-     units                = days since 1950-01-01 00:00:00
-[...]
+     units                = degree Celsius
+     scale_factor         = 10
+     comments             = this is a string attribute with Unicode Ω ∈ ∑ ∫ f(x) dx
+
+Global attributes
+  title                = this is a test file
 ```
 
 ## Load a netCDF file
@@ -89,7 +141,7 @@ Loading a variable with known structure can be achieved by accessing the variabl
 
 ```julia
 # The mode "r" stands for read-only. The mode "r" is the default mode and the parameter can be omitted.
-ds = NCDataset("/tmp/test.nc","r")
[…]
