### 2024-03-20 #6504 @Alexander-Barth
Thanks @boriskaus for this good suggestion! I implemented this change here: https://github.com/Alexander-Barth/NCDatasets.jl/commit/2fe4efb9dd7328cc0801bccd376e3dd950908819 .
