### 2024-03-21 #6504 @boriskaus
@Alexander-Barth as part of the review they ask me to verify that your performance claims are correct. You did a nice job of describing how to set this up and run it, but unfortunately I don’t have access to a Linux machine where I have root access. The testing doesn’t seem to work on a Mac.
Is there any possibility for this to run in a way that doesn’t require root access? I guess others may run into the same issue if trying this.
