### 2024-03-21 #6504 @boriskaus
Yes that seems a fine idea to me
