### 2024-03-21 27a2992
make the caches droping optional
