+++ test/perf/benchmark.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
-julia   benchmark-julia-NCDatasets.jl
-python3 benchmark-python-netCDF4.py
-Rscript benchmark-R-ncdf4.R
+args="$@"
+julia   benchmark-julia-NCDatasets.jl $args
+python3 benchmark-python-netCDF4.py $args
+Rscript benchmark-R-ncdf4.R $args
 
 julia summary.jl
