### 2024-03-22 #6504 @Alexander-Barth
I updated the benchmark here: https://github.com/Alexander-Barth/NCDatasets.jl/commit/27a29922b282e097cb230d8e4b5b7eeb8c6cdb7f

In fact, "not dropping the caches" is now the default so that you should be able to reproduce the benchmark simply with running the script `./benchmark.sh`. I also updated the description here:

https://github.com/Alexander-Barth/NCDatasets.jl/blob/master/test/perf/README.md
