### 2024-03-22 #6504 @boriskaus
thanks a lot. I managed to reproduce this on Mac (M2). The numbers are fairly similar and Julia remains on top:

| Module           | median | minimum |  mean | std. dev. |
|:---------------- | ------:| -------:| -----:| ---------:|
| R-ncdf4          |  0.341 |   0.304 | 0.354 |     0.035 |
| python-netCDF4   |  0.225 |   0.201 | 0.227 |     0.018 |
| julia-NCDatasets |  0.219 |   0.205 | 0.220 |     0.009 |
