### 2024-04-09 8f452ad
Throw informative error for unsupported vtype
Currently, defVar fails with something like
```
ERROR: KeyError: key DateTime not found
```
This PR throws a more informative error message
