+++ test/test_writevar.jl
@@ -12,6 +12,8 @@ ds = NCDataset(filename,"c")
 defDim(ds,"lon",sz[1])
 defDim(ds,"lat",sz[2])
 
+# vartype not supported
+@test_throws ErrorException defVar(ds,"var-DT",DateTime,("lon","lat"))
 
 # variables
 for T in [UInt8,Int8,UInt16,Int16,UInt32,Int32,UInt64,Int64,Float32,Float64]
