### 2024-04-18 #6504 @lanari
@Alexander-Barth, please accept my apologies for the delay in reviewing your work.

@majensen This is a well-designed package and I think it will be very useful for the community. The repository contains all the necessary files and the examples given are easy to reproduce. The software paper is clear and well written. I strongly recommend publication.

For comparison, here is the package's performance on a MacBook Pro 2.3GHz 8-core Intel Core i9:

|  Module |  median | minimum  | mean  | std. dev.  |
|---|---|---|---|---|
| julia-NCDatasets  | 0.312  | 0.247  |  0.384 |  0.030 |
