+++ Project.toml
@@ -20,7 +20,7 @@ CFTime = "0.1.1"
 CommonDataModel = "0.3.4"
 DataStructures = "0.17, 0.18"
 DiskArrays = "0.3.22"
-NetCDF_jll = "=400.701.400, =400.702.400, =400.902.5, =400.902.208, =400.902.209"
+NetCDF_jll = "=400.701.400, =400.702.400, =400.902.5, =400.902.208, =400.902.209, =400.902.211"
 NetworkOptions = "1.2"
 julia = "1.6"
 
