### 2024-04-23 6735b42
try to save attributes of the type Vector{Any}
