+++ src/variable.jl
@@ -335,6 +335,8 @@ end
 export fillvalue
 
 
+access() = error("This function is available after importing MPI")
+
 """
     a = nomissing(da)
 
