### 2024-04-23 cf65b9a
weak dependencies require julia 1.9
