+++ .github/workflows/ci.yml
@@ -10,9 +10,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.6'
+          - '1.9'
           - '1'
-          #- '~1.8.0-0'
           - 'nightly'
         os:
           - ubuntu-latest
