+++ Project.toml
@@ -29,7 +29,7 @@ DiskArrays = "0.3.22"
 MPI = "0.20"
 NetCDF_jll = "=400.701.400, =400.702.400, =400.902.5, =400.902.208, =400.902.209, =400.902.211"
 NetworkOptions = "1.2"
-julia = "1.6"
+julia = "1.9"
 
 [extras]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
