### 2024-04-23 deb6950
enable MPI test
