+++ Project.toml
@@ -27,7 +27,7 @@ CommonDataModel = "0.3.4"
 DataStructures = "0.17, 0.18"
 DiskArrays = "0.3.22"
 MPI = "0.20"
-NetCDF_jll = "=400.701.400, =400.702.400, =400.902.5, =400.902.208, =400.902.209, =400.902.211"
+NetCDF_jll = "=400.902.211"
 NetworkOptions = "1.2"
 julia = "1.9"
 
