### 2024-04-23 cc51bab
add MPI as test dependency
