+++ Project.toml
@@ -30,13 +30,3 @@ MPI = "0.20"
 NetCDF_jll = "=400.902.211"
 NetworkOptions = "1.2"
 julia = "1.9"
-
-[extras]
-Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Dates", "Test", "Random", "Printf", "IntervalSets"]
