### 2024-04-24 e792afa
skip MPI test for windows (issue #122)
