### 2024-04-24 e61126b
use Int32 in the example
