### 2024-04-24 96990c4
add docs for MPI support
