### 2024-04-25 d792d8a
change parallel access to whole dataset (issue #122, thanks @pgf)
