+++ ext/NCDatasetsMPIExt.jl
@@ -16,7 +16,7 @@ using NCDatasets:
 
 import NCDatasets:
     NCDataset,
-    access
+    paraccess
 
 function nc_create_par(path,cmode::Integer,mpi_comm,mpi_info)
     ncidp = Ref(Cint(0))
@@ -53,14 +53,14 @@ function parallel_access_mode(par_access::Symbol)
 end
 
 """
-    NCDatasets.access(ncv::Variable,par_access::Symbol)
-    NCDatasets.access(ds::NCDataset,par_access::Symbol)
+    NCDatasets.paraccess(ncv::Variable,par_access::Symbol)
+    NCDatasets.paraccess(ds::NCDataset,par_access::Symbol)
 
-Change the parallel access mode of the variable `ncv` or all variables of the dataset `ds` for writing or reading data. `par_access` is either `:collective` or `:independent`. `NCDatasets.access` will raise an error if `MPI` is not loaded.
+Change the parallel access mode of the variable `ncv` or all variables of the dataset `ds` for writing or reading data. `par_access` is either `:collective` or `:independent`. `NCDatasets.paraccess` will raise an error if `MPI` is not loaded.
 
 More information is available in the [NetCDF documentation](https://web.archive.org/web/20240414204638/https://docs.unidata.ucar.edu/netcdf-c/current/parallel_io.html).
 """
-function access(ncv::Variable,par_access::Symbol)
+function paraccess(ncv::Variable,par_access::Symbol)
     ds = dataset(ncv)
     ncid = ds.ncid
     varid = ncv.varid
@@ -68,15 +68,15 @@ function access(ncv::Variable,par_access::Symbol)
     if nc_inq_format(ncid) in (NC_FORMAT_NETCDF4, NC_FORMAT_NETCDF4_CLASSIC)
         nc_var_par_access(ncid,varid,parallel_access_mode(par_access))
     else
-        error("The netCDF 3 and 5 formats do not allow different access methods per variable. You need to call this function for the whole data set: NCDatasets.access(ds,$par_access)")
+        error("The netCDF 3 and 5 formats do not allow different access methods per variable. You need to call this function for the whole data set: NCDatasets.paraccess(ds,$par_access)")
     end
 end
 
 # set collective or independent IO globally (for all variables)
-function access(ds::NCDataset,par_access::Symbol)
+function paraccess(ds::NCDataset,par_access::Symbol)
     if nc_inq_format(ds.ncid) in (NC_FORMAT_NETCDF4, NC_FORMAT_NETCDF4_CLASSIC)
         for (varname,ncv) in ds
-            access(ncv.var,par_access)
+            paraccess(ncv.var,par_access)
         end
     else
         # only for PnetCDF
