### 2024-04-25 a9da4ae
add MPI to docs/make.jl
