### 2024-04-25 34c0d32
run tests in test env. for ci-binary-builder
