### 2024-04-25 4348c09
set project for mpiexec
