### 2024-04-25 #6504 @Alexander-Barth
I just created a new release with the Zenodo-Github link activated:

The DOI is [10.5281/zenodo.11067062](https://doi.org/10.5281/zenodo.11067062) corresponding to the NCDatasets version v0.14.4.
