### 2024-04-25 0768d5c
feat: type-stable `variable`
