### 2024-04-25 084329c
supply errors for @test_throws
