+++ test/test_variable.jl
@@ -325,10 +325,10 @@ var1 = variable(ds, "temperature", Float32, ("lon", "lat"))[1:5, 1:10]
 var2 = variable(ds, "temperature", Float32, ("lat", "lon"))[1:5, 1:10]  # TODO: Ensure this errors
 @test var1 == var2
 # ... but incorrect order of `dimnames` and out-of-bounds slices errors
-@test_throws variable(ds, "temperature", Float32, ("lat", "lon"))[:, 1]  # `:` is actually `1:110`
+@test_throws MethodError variable(ds, "temperature", Float32, ("lat", "lon"))[:, 1]  # `:` is actually `1:110`
 # incorrect `dimnames` errors
-@test_throws variable(ds, "temperature", Float32, ("lon", "lat", "time"))
-@test_throws variable(ds, "temperature", Float32, ("lon",))
+@test_throws NCDatasets.NetCDFError variable(ds, "temperature", Float32, ("lon", "lat", "time"))
+@test_throws MethodError variable(ds, "temperature", Float32, ("lon",))
 # Incorrect `DType` doesn't error, but gives incorrect result
 d1 = variable(ds, "temperature", Float32, ("lon", "lat"))[1:5, 1:3]
 d2 = variable(ds, "temperature", Float64, ("lon", "lat"))[1:5, 1:3]
