### 2024-05-28 cf722ab
test NCDatasets.nc_inq_var
