+++ test/test_lowlevel.jl
@@ -57,6 +57,12 @@ for sampledata in samples
     xtype2 = NCDatasets.nc_inq_vartype(ncid,varid)
     @test xtype == xtype
 
+    name2,jltype2,dimids2,natts2 = NCDatasets.nc_inq_var(ncid,varid)
+    @test name2 == varname
+    @test jltype2 == T
+    @test dimids2 == reverse(dimids)
+    @test natts2 == 1
+
     attrval = NCDatasets.nc_get_att(ncid, varid, "attr-string-list")
     @test attrval == ["one","two"]
 
