### 2024-05-28 4355ba2
test set/get netcdf rc configuration
