### 2024-05-28 80430f7
test NCDatasets.nc_inq_filter_avail
