### 2024-05-28 7ee06b6
test nc_get_var1
