### 2024-05-28 071be21
test nc_put_var1
