### 2024-05-28 #6504 @kthyng
@Alexander-Barth Can you update your Zenodo title and author list to exactly match your JOSS paper? This is a preference not a requirement so the JOSS archive is self-consistent.
