### 2024-05-28 #6504 @kthyng
@Alexander-Barth Also please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
