+++ paper/paper.bib
@@ -1,10 +1,10 @@
 @manual{OGC_netCDF,
-  organization  = "Open Geospatial Consortium",
-  title         = "NetCDF Binary Encoding Extension Standard: NetCDF Classic and 64-bit Offset Format",
+  organization  = {Open Geospatial Consortium},
+  title         = {{NetCDF Binary Encoding Extension Standard: NetCDF Classic and 64-bit Offset Format}},
   year          =  2011,
   month         =  4,
-  number        = "OGC 10-092r3",
-  url          = "http://www.opengis.net/doc/IS/netcdf-binary/1.0",
+  number        = {OGC 10-092r3},
+  url          = {http://www.opengis.net/doc/IS/netcdf-binary/1.0},
 }
 
 @inproceedings{Rew2006,
@@ -12,13 +12,13 @@
   year = {2006},
   month = {01},
   organization = {Conference: 22nd International Conference on Interactive Information Processing Systems for Meteorology, Oceanography, and Hydrology},
-  title = {NetCDF-4: Software Implementing an Enhanced Data Model for the Geosciences}
+  title = {{NetCDF-4: Software Implementing an Enhanced Data Model for the Geosciences}}
 }
 
 @article{Rew90,
   author={Rew, R. and Davis, G.},
   journal={IEEE Computer Graphics and Applications},
-  title={NetCDF: an interface for scientific data access},
+  title={{NetCDF: an interface for scientific data access}},
   year={1990},
   volume={10},
   number={4},
@@ -27,12 +27,12 @@
 }
 
 @manual{OGC_Zarr,
-  organization  = "Open Geospatial Consortium",
-  title         = "Zarr Storage Specification 2.0 Community Standard",
+  organization  = {Open Geospatial Consortium},
+  title         = {{Zarr Storage Specification 2.0 Community Standard}},
   year          =  2022,
   month         =  6,
-  number        = "21-050r1",
-  url          = "http://www.opengis.net/doc/CS/zarr/2.0"
+  number        = {21-050r1},
+  url          = {http://www.opengis.net/doc/CS/zarr/2.0}
 }
 
 
@@ -47,7 +47,7 @@
 
 @article{Doglioni2023,
   AUTHOR = {Doglioni, F. and Ricker, R. and Rabe, B. and Barth, A. and Troupin, C. and Kanzow, T.},
-  TITLE = {Sea surface height anomaly and geostrophic current velocity from altimetry measurements over the Arctic Ocean (2011--2020)},
+  TITLE = {{Sea surface height anomaly and geostrophic current velocity from altimetry measurements over the Arctic Ocean (2011--2020)}},
   JOURNAL = {Earth System Science Data},
   VOLUME = {15},
   YEAR = {2023},
@@ -58,7 +58,7 @@
 
 @article{Belgacem21,
   AUTHOR = {Belgacem, M. and Schroeder, K. and Barth, A. and Troupin, C. and Pavoni, B. and Raimbault, P. and Garcia, N. and Borghini, M. and Chiggiato, J.},
-  TITLE = {Climatological distribution of dissolved inorganic nutrients in the western Mediterranean Sea (1981--2017)},
+  TITLE = {{Climatological distribution of dissolved inorganic nutrients in the western Mediterranean Sea (1981--2017)}},
   JOURNAL = {Earth System Science Data},
   VOLUME = {13},
   YEAR = {2021},
@@ -83,7 +83,7 @@
 
 @ARTICLE{Shahzadi21,
   AUTHOR={Shahzadi, K. and Pinardi, N. and Barth, A. and Troupin, C. and Lyubartsev, V. and Simoncelli, S.},
-  TITLE={A New Global Ocean Climatology},
[…]
