### 2024-05-28 a007ff1
fix imprecise formulation
