### 2024-05-28 #6504 @Alexander-Barth
Thanks a lot for spotting these issues! I fixed the capitalization in the references, and updated author/title of the Zenodo archive to match the JOSS paper.

While proof-reading the paper, I made these (small) changes to the paper:

https://github.com/Alexander-Barth/NCDatasets.jl/commit/a007ff1e28224edc5bcde99e5c48eda75c600168

https://github.com/Alexander-Barth/NCDatasets.jl/commit/f8d7643879fc1f828e804266655df09138dd4281

I hope that this is ok for you :-)
