### 2024-05-29 #6504 @kthyng
Congratulations on your new publication @Alexander-Barth! Many thanks to @majensen and to reviewers @lanari and @boriskaus for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
