### 2024-05-29 #6504 @kthyng
@Alexander-Barth If you're interested in joining JOSS as a reviewer, please sign up in the database here: https://reviewers.joss.theoj.org/.
